[gp-cli] move cmd files
