+++ components/gitpod-cli/cmd/tasks.go
@@ -5,7 +5,6 @@
 package cmd
 
 import (
-	"github.com/gitpod-io/gitpod/gitpod-cli/cmd/tasks"
 	"github.com/spf13/cobra"
 )
 
@@ -20,32 +19,6 @@ var tasksCmd = &cobra.Command{
 	},
 }
 
-var attachTaskCmdOpts struct {
-	Interactive bool
-	ForceResize bool
-}
-
-// listTasksCmd represents the tasks list command
-var listTasksCmd = &cobra.Command{
-	Use:   "list",
-	Short: "Lists the workspace tasks and their state",
-	Run:   tasks.ListTasksCmd,
-}
-
-// attachTaskCmd represents the attach task command
-var attachTaskCmd = &cobra.Command{
-	Use:   "attach <id>",
-	Short: "Attach to a workspace task",
-	Args:  cobra.MaximumNArgs(1),
-	Run:   tasks.AttachTasksCmd,
-}
-
 func init() {
 	rootCmd.AddCommand(tasksCmd)
-
-	tasksCmd.AddCommand(listTasksCmd)
-	tasksCmd.AddCommand(attachTaskCmd)
-
-	attachTaskCmd.Flags().BoolVarP(&attachTaskCmdOpts.Interactive, "interactive", "i", true, "assume control over the terminal")
-	attachTaskCmd.Flags().BoolVarP(&attachTaskCmdOpts.ForceResize, "force-resize", "r", true, "force this terminal's size irregardless of other clients")
 }
